Move LooksLikeGarbage from herald to internal/ai
LooksLikeGarbage validates AI summary output, so internal/ai is its natural
home — and the move is a prerequisite for the staged pipeline (#113): the
summarize stage lives in internal/pipeline, which cannot import the root
herald package (engine.go will import pipeline once it delegates to it, so the
dependency must point one way). Pure relocation: the function and its test move
verbatim; the three call sites (engine.go, screen.go, main.go) switch to
ai.LooksLikeGarbage, and screen.go drops its now-unused herald import.
